backend/afro-storm-pipeline/requirements.txt
# AFRO Storm Intelligence Pipeline
# Multi-modal AI-powered climate & health surveillance

# Core Data Processing
numpy==1.26.4
pandas==2.2.0
xarray==2024.1.0
netCDF4==1.7.1
scipy==1.12.0

# Geospatial
geopandas==0.14.3
shapely==2.0.3
rasterio==1.3.9
pyproj==3.6.1
folium==0.15.1
geopy==2.4.1

# Climate Data APIs
cdsapi==0.6.1  # Copernicus Climate Data Store
ecmwf-api-client==1.6.3
ecmwf-opendata>=0.3.26  # ECMWF open data (free 0.25° forecasts)
requests==2.31.0
aiohttp==3.9.3

# AI/ML
torch>=2.2.0
transformers>=4.37.2
anthropic>=0.18.1
openai>=1.12.0
langchain>=0.1.6
# chromadb>=0.4.22

# Computer Vision
opencv-python==4.9.0.80
pillow==10.2.0

# Time Series
prophet==1.1.5
statsmodels==0.14.1

# Graph Database
neo4j==5.17.0
py2neo==2021.2.4

# Web Scraping & APIs
beautifulsoup4==4.12.3
selenium==4.17.2
httpx>=0.25.2,<1.0.0

# Messaging & Alerts
twilio==8.12.0
python-telegram-bot==20.7
discord.py==2.3.2
slack-sdk==3.26.2

# Task Scheduling
apscheduler==3.10.4
celery==5.3.6
redis==5.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
loguru==0.7.2
tqdm==4.66.1

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
python-multipart==0.0.9

# MoStar Grid - Local AI via Ollama
ollama==0.1.7

# MoStar Grid - Knowledge Graph (already have neo4j)
# neo4j==5.17.0  # Already listed above

# Visualization
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
